Bumped everything except Kusto and split the Kusto tests for smoother run isolation
