+++ src/Eshopworld.Telemetry/Eshopworld.Telemetry.csproj
@@ -26,11 +26,11 @@
 
   <ItemGroup>
     <PackageReference Include="Eshopworld.Core" Version="2.0.5" />
-    <PackageReference Include="Microsoft.ApplicationInsights" Version="2.8.1" />
+    <PackageReference Include="Microsoft.ApplicationInsights" Version="2.9.1" />
     <PackageReference Include="Microsoft.Azure.Kusto.Ingest.NETStandard" Version="4.0.3" />
     <PackageReference Include="Microsoft.Azure.Services.AppAuthentication" Version="1.0.3" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
-    <PackageReference Include="System.Reactive" Version="4.1.2" />
+    <PackageReference Include="System.Reactive" Version="4.1.3" />
   </ItemGroup>
 
 </Project>
